% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/thezone.R
\name{enterthezone}
\alias{enterthezone}
\title{Live in the zone}
\usage{
Set the following parameters:
rediska.channel - the name of the list to which other processes should lpush.
rediska.msg_threshold - maximum number of messages to be shown. If there are more messages a warning will be shown.
rediska.exitthezone - name of the variable through which one can signal rediska to exit the zone.
}
\description{
This function adds a callback to every task that will read messages from
a redis list and write them to your R console. Useful to pipe notifications
from other processes and stay in the zone.
}

