% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/commands.R
\name{raw_command}
\alias{raw_command}
\title{Execute raw redis command}
\usage{
raw_command(command)
}
\arguments{
\item{command}{character. The command string that sould be executed}
}
\value{
The response from redis-cli
}
\description{
Execute raw redis command
}
\examples{
\dontrun{
  rediska:::raw_command("PING")
  # [1] "PONG"
}
}

