% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/commands.R
\docType{data}
\name{rediska}
\alias{rediska}
\title{An environment that contains methods for reading from / writing to redis}
\format{\preformatted{<environment: 0x7f9d11114cd0> 
}}
\usage{
rediska$ping()
rediska$set("foo", "bar")
rediska$set(c("bar", "baz"))
rediska$get("foo")

# You can also use
rediska$raw_command("SET foo bar")
}
\description{
You can see the list of all available functions by executing \code{rediska:::command_names}.
}
\keyword{datasets}

